Usage of index template for settings and mapping
Synonyms are still defined at the creation of the index and will
override the defined in the templates. But at least one synonym need to
be defined on the template.

Elasticsearch templates has to be initialised before importing data,
this is done with the `mimir_init` binary.

This PR may change the result of geocoder-tester since a lot of indexes
will have less shard than previously and that scoring is done per shard.
Having less shard reduce load on query and recent version of
Elasticsearch have a limit on the number of shard that can be queried.

Recommanded values are the following:
  - less 3M documents: 1 shard
  - between 3M and 5M documents with an expected growth over 5M: 2 shards.
  - More than 5M: int (number of expected documents / 5M +1)

With this PR each index has mapping defined for only one type, this is
required by recent version of Elasticsearch.