add an abstraction over weight
The Weight is an object used to for the Mimir's Object weight.

A Weight's value must be a float between 0 and 1.

Since we need interior mutability for the weight and we need it to be
std::sync, it emulates an AtomicFloat.
The main idea is to store the value in an AtomicUSize and convert it to
a float with the `get_val` method.

The weight needs to be normalized before used.
The normalization works by giving the maximum value to the normalize
method, this way we can compute a weight between 0 and 1.