dont take label's size into account
As the label (and the full_label) are generated fields, we don't want to
take the label's size into account when scoring.

without this , for the query `saint raphael`

`Saint-Raphaël (24160), Dordogne, Nouvelle-Aquitaine, France`

had a better score than

`Saint-Raphaël (83530-83700), Var, Provence-Alpes-Côte d'Azur, France`

even if the second one had a bigger score, because since the first label
is shorter, we matched a bigger portion of it.