add a dockerfile for the import bin
this way we'll be able to push some mimir images in hub.docker.com and
run mimir's job in container.

We might want to rename `Dockerfile` to `Docker_bragi` but it will
likely break some Kisio's integration, so I think we should discuss it.

If that's ok for you I think we should set up some automated build in
dockerhub. It might be better if it's in the kisio's organisation but if
you don't want to do it, I can set it up in qwantreasearch's
organisation.

the remaining question is how we're gonna name it.
* `bragi` ? `mimir_bragi` ?
* `mimir_import` ?