do not index useless fields
by default elasticsearch index all unkown field (not in the mapping).

this means that it will take some time and some space to index some
useless field. Like for example the big admins name (thrice in the poi
index :wink: )

https://www.elastic.co/guide/en/elasticsearch/guide/current/dynamic-mapping.html

adding `"dynamic": "false"` to the mapping tells ES not to index the
unkown field (but they are still stored).