fix useless method
A test method was never called (my bad :confused: ).

The thing is that the test was not passing either, because if we look
for a not existing index, we won't have a specific error as the test
state, but a generic one (no object found).

This is done this way because the rubber `get_indexes` method always
return a non empty list of elements (there is always at least
"munin_geo_data" returned)

I think the get_indexes must also be refactored, but it will be for
another PR.