add a default timeout
add a cli param `default-es-timeout` (env var
`BRAGI_DEFAULT_ES_TIMEOUT`) to set the default timeout to ES connections
(in ms)

If present we'll take the existing query `timeout` parameter, else the
default one.

for information the current error when hitting the timeout is:

```
HTTP/1.1 503 Service Unavailable
Access-Control-Allow-Origin: *
Content-Type: application/json
{
    "long": "invalid query Resource temporarily unavailable (os error
11)",
    "short": "query error"
}

```

A bit cryptic, but I think that'll do for the moment.

I don't see an easy way to test this.