Fix labels normalization
We don't want to use the size of a label since it is generated.
It has been implemented in
https://github.com/CanalTP/mimirsbrunn/pull/229 but I forgot to also
remove the norms for the full field (we want to boost a response when we
match entire tokens.

The test case for us was that the query `Lyon` was having the results:

```
Lyon, Coahoma County, Mississippi, United States <---- small village in
mississipi
Lyon, Métropole de Lyon, Auvergne-Rhône-Alpes, France <---- big french
city with a lot of population
```

with this PR, we now have the french city first