Localized labels
use  [opencage's
configuration](https://github.com/CanalTP/address-formatter-rs) to have
a localized label.

This is still a work in progress since for the moment this changes all
the labels, we need to see if that's ok for us or not.

Note: I also removed serde_derive as stated in the serde's documentation
since I had some strange behavior with rs-es serde's imports.