Output the stop's lines
When stops have been imported with a ntfs (and not only a gtfs's
stops.txt file) we store the stop's lines to output them in bragi.

This will makes it possible to add a nice picto in the search results
with the lines passing by the stop. It will makes it easier to
differentiate the results at first glance.

Note: the lines are not indexed by ES, so this should not change any
results.