First shot at a revamped documentation
follow https://github.com/CanalTP/mimirsbrunn/issues/231

Revamp a bit the documentation.

There is still a lot of TODOs, and I don't plan of addressing them all
in this PR, but this should be a better starting point for someone
discovering this repository.

Feel free to tell me if some parts are difficult to understand or not
correctly organized.