Feature: Handling private indices
This commit adds a flag 'private' to rubber's `publish_index` function
so that a private index is not aliased by global indices, eg munin.

This flag is propagated to poi2mimir's command line, so that we can
import private POIs.

There are also modifications of the elastic search wrapper to allow
queries on non global indices.