Feature: Add explain to bragi
This change adds a query url parameter to bragi (`debug=true`),
which triggers a so called explain query in elasticsearch, and
forwards the results back as json.

By default, when the parameter is not present, the flag debug is false.

This new feature can be used to analyze the score assigned by
elasticsearch to a result.

This change adds an 'explain' object to a geocodejson feature object,
which is not part of the initial specification.

This change addresses issue #214