WIP Replace gst with rstar; Update dependencies.
gst is unmaintained (4 years) and needs to be replaced.
rstar is the obvious contender for replacing gst.

By making use of geo algorithm, this change simplifies the code
for the AdminGeoFinder.

This proposed change has a breaking consequence however. We can no
longer insert admins with boundaries from 'osm2mimir'. We can still
do with 'cosmogony and cosmogony2mimir'. Here is why:

In the AdminGeoFinder, we strip the boundary from the Admin, to save
memory. In 'osm2mimir', we store the admins found in the pbf file in
the AdminGeoFinder, and then iterate over them to index them in
Elasticsearch. With cosmogony2mimir, we simply iterate over all the
admins found in the cosmogony file, and index them (we don't use the
AdminGeoFinder).