Fix dropped fb-file option
When the db-file option is used with osm2mimir, but the binary wasn't
compiled with the db-storage feature, the option was silently dropped.

The user now has a warning:
WARN You are trying to use DB storage but the program wasn't compiled with the feature 'db-storage'.
WARN You should either recompile with that feature, or use in memory storage.
ERRO Unable to use DB Storage for OSM Store without feature turned on