[FIX]: Display error in context
This is a stopgap fix to retrieve the error context in some situation
where it was dropped and not displayed to the user. This may not be
the way failure was meant to be used (rather using error chains), but
this solution gets the error context to the end user. A long term
solution will involve changing to eg anyhow, snafu or other error
libraries