[Changed] Add negative flags
We introduce 3 command line argument to let the user specify that
he/she does not want to import a certain kind of place. Prior to this,
the absence of the flag was meant to signify the user did not want to
import a certain kind of place, which interfered with the configuration.