bragi dsl: fix ranking and i18n (#729)
* bragi dsl: some readability changes

* bragi dsl: fix typo in localized fields

* bragi dsl: smaller global weights

Since ES2, the default similarity algorithm has changed (and the older
one is no longer an option). This new algorithm provides bigger weights
than before which leads to inconsistent contribution between string
query, type query and weight/proximity boosts.

A quick analysis on the older version of Bragi running ES2 shows that
weights were normalized by a factor of 1/1200, so this change multiplies
global weights by 1/1200 to get more consistent values.

* bragi dsl: fix admin scoring for non-admins

By default when there is no function matching in a function_score the
overall score is 1. In `build_admin_weight_query` this meant that we add
a score of 1 to non-admin documents, which was a huge boost as most scores
are below 1.

* query dsl: some cosmetic changes

* bragi dsl: most clones can be avoided in dsl.rs