fix non-string parameters for autocomplete-explain endpoint
Due to a limitation in serde_qs, you need to explicitly specify how you
parse a field after a layer of #[serde(flatten)], this is what is
already done for some existing fields in the autocomplete parameters.

However I don't think this workarround would work for two layers of
flatten, so the simplest fix is probably to repeat the parameters
definition.