doc updates: better support for dot-separated keys
For some reason when using a dot-sparated key chain to update a document
value, elasticsearch will attempt to update the mapping. Which conflicts
with `properties` which is flattened.

For example the following update operation fails:

```json
{
   "script": {
       "source": "ctx._source['properties.image'] = \"test\""
    }
}
```

> Could not dynamically add mapping for field [properties.image]. Existing mapping for [properties] must be of type object but found [flattened].

And the following one succeeds

```json
{
   "script": {
       "source": "ctx._source['properties']['image'] = \"test\""
    }
}
```

So the approach of this PR is to manually split key and subkeys and
generate the subtree for ES to update:

```json
{
    "doc": {
        "properties": {
            "image": "test"
        }
    }
}
```