updates: now taking a list of updates for each document
It turns out that ES can't easily handle the cases were several updates
to the same document are submitted concurrently, this results in the
following message: "Error: [doc id]: version conflict".

An easy way to avoid this issue is to generate a single operation per
document, it could intuitively also help with performances.